refactor(trade): split files exceeding 700-line hard cap + stale doc fix
CodeRabbit's second-round review on PR #438 flagged 4 findings; the
3 file-size splits resolve the CLAUDE.md §"File organization" hard
cap, and the stale doc comment is updated.

execute.rs (1307 lines) → execute/ directory:
- mod.rs (501): public entry (`handle_execute_trade`) + TradeSide +
  TradeFinalBalances + TradeAbort enum + Display/From + the
  `trade_abort_to_results_codes` mapping + outbound wire helpers
  (`send_results_to_both`, `send_on_trade_results`).
- swap.rs (471): atomic-swap transaction internals
  (`atomic_swap` + advisory lock + `lock_items` + slot reservation
  + parking sentinel + the two-phase row mover). Internal helpers
  exposed as `pub(super)` for the unit tests.
- tests.rs (350): 4 cfg-only unit test modules
  (advisory-lock alignment, slot-exclusion accounting, parking
  sentinel distinctness, asymmetric ETradeResults mapping).

tests.rs (live-DB integration, 976 lines) → tests/ directory:
- mod.rs (207): shared helpers (cleanup, insert_account_and_player,
  insert_item, naquadah_of, owner_of, slot_id_of, TradeFixtures,
  fixtures, make_state, tradeable_type_ids) — all `pub(super)`
  for the test submodules.
- commit.rs (393): core commit/rollback (happy path + 6 rollback
  variants).
- container_whitelist.rs (180): the INV_MAIN-only whitelist
  regression guards + INV_MAIN positive control.
- no_db.rs (123): no-DB-pool early-return + check-ordering tests.
- slot_reservation.rs (150): full-bag-swap-while-trading-slot-away
  regression guard.

tests/handlers.rs (cell trade handler tests, 719 lines) → 2 files:
- handlers.rs (334): happy-path + session lifecycle
  (open / cancel / disconnect / range / QA workaround).
- handlers_validation.rs (411): validation rejects
  (version mismatch, self-trade, hijack guards, NPC caller/partner,
  proposal dedup, truncated lock-state args).

Also addresses CodeRabbit's stale doc comment on `TradeAbort`:
the comment said "future work can map these variants to asymmetric
ETradeResults codes" but `trade_abort_to_results_codes` already does
this; updated to point at the mapping function.

No production-behavior change. All 53 trade-related unit tests pass;
all 12 live-DB integration tests pass when `DATABASE_URL` is set.
